Rename authentication directory
